services/_dash/src/pyproject.toml
[tool.poetry]
name = "plotly-dashboards"
version = "0.1.0"
description = "Dashboard server"
authors = ["QR Space <qrspace.io>"]
license = "All Rights Reserved"

[tool.poetry.dependencies]
python = "^3.10"
Flask = "^2.0.2"
Flask-Cors = "^3.0.10"
Flask-Session = "^0.4.0"
flask-talisman = "^0.8.1"
Werkzeug = "^2.0.2"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.3"
requests = "^2.27.0"
loguru = "^0.6.0"
pydantic = "^1.9.0"
redis = "^4.1.0"
dash = "^2.5.1"
pandas = "^1.3.5"
python-dateutil = "^2.8.2"
Bootstrap-Flask = "^2.0.2"
dash-bootstrap-components = "^1.2.0"
python-dotenv = "^0.20.0"
Authlib = "^1.0.1"
swifter = "^1.2.0"
dash-mantine-components = "^0.10.2"
sd-material-ui = "^4.6.0"
dash-daq = "^0.5.0"
dash-tabulator = "^0.4.2"
dash-extensions = "^0.1.6"


[tool.poetry.dev-dependencies]
pytest = "^6.2.5"
black = "^22.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::DeprecationWarning",
]